refactor: 예외 클래스 및 핸들러 패키지 구조 변경 (#45)
