+++ src/main/kotlin/kr/wooco/woocobe/common/domain/exception/BaseErrorCode.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.common.exception
+package kr.wooco.woocobe.common.domain.exception
 
 import org.springframework.http.HttpStatus
 
@@ -10,6 +10,5 @@ enum class BaseErrorCode(
     METHOD_ARGUMENT_TYPE_MISMATCH("METHOD_ARGUMENT_TYPE_MISMATCH", "method argument error", HttpStatus.BAD_REQUEST),
     METHOD_NOT_SUPPORT("METHOD_NOT_SUPPORTED", "method not supported", HttpStatus.METHOD_NOT_ALLOWED),
     NO_RESOURCE("NO_RESOURCE", "no such resource", HttpStatus.NOT_FOUND),
-
     UNKNOWN("UN_KNOWN_ERROR", "not included in any exception class", HttpStatus.INTERNAL_SERVER_ERROR),
 }
